requirements.txt
langchain==0.0.350
openai==1.3.8
streamlit==1.28.2
python-dotenv==1.0.0